# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptOut": {
        "properties": {
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttemptItem"
            },
            "title": "Items",
            "type": "array"
          },
          "mode": {
            "enum": [
              "normal",
              "exam"
            ],
            "title": "Mode",
            "type": "string"
          },
          "quiz": {
            "$ref": "#/components/schemas/QuizOut"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "enum": [
              "in_progress",
              "marking",
              "marked"
            ],
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttemptSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "quiz",
          "mode",
          "status",
          "started_at",
          "deadline",
          "submitted_at",
          "items"
        ],
        "title": "AttemptOut",
        "type": "object"
      },
      "AttemptStarted": {
        "properties": {
          "attempt_id": {
            "format": "uuid",
            "title": "Attempt Id",
            "type": "string"
          },
          "quiz": {
            "$ref": "#/components/schemas/QuizOut"
          }
        },
        "required": [
          "quiz",
          "attempt_id"
        ],
        "title": "AttemptStarted",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScoreOverride": {
        "additionalProperties": false,
        "properties": {
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "score"
        ],
        "title": "ScoreOverride",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Revision OS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/answers/{answer_id}/dispute": {
      "post": {
        "description": "Re-mark a Claude-marked answer with the stronger model.",
        "operationId": "dispute_mark_api_v1_answers__answer_id__dispute_post",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dispute Mark",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/answers/{answer_id}/override": {
      "post": {
        "description": "Set the mark yourself (the original is kept for the record).",
        "operationId": "override_mark_api_v1_answers__answer_id__override_post",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreOverride"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Override Mark",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/attempts": {
      "get": {
        "description": "Your quizzes and mock exams in a module, newest first.",
        "operationId": "list_attempts_api_v1_attempts_get",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttemptListItem"
                  },
                  "title": "Response List Attempts Api V1 Attempts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Attempts",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}": {
      "get": {
        "description": "The quiz as you answer it (no answers shown), or your results once\nsubmitted. An exam past its deadline is submitted automatically.",
        "operationId": "get_attempt_api_v1_attempts__attempt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Attempt",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/responses/{question_id}": {
      "put": {
        "operationId": "save_response_api_v1_attempts__attempt_id__responses__question_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseSave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Response",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/responses/{question_id}/photo": {
      "post": {
        "description": "A photo of your handwritten working (raw request body). Returns\nClaude's transcription for you to check and edit into your answer.",
        "operationId": "upload_photo_api_v1_attempts__attempt_id__responses__question_id__photo_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoTranscription"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Photo",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/submit": {
      "post": {
        "description": "Submit the whole quiz. Exact marks appear at once; Claude's marks and\nexplanations follow (status \"marking\", then \"marked\").",
        "operationId": "submit_attempt_api_v1_attempts__attempt_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Attempt",
        "tags": [
          "practice"
        ]
      }
    },
...
    "/api/v1/flashcards": {
      "get": {
        "operationId": "list_flashcards_api_v1_flashcards_get",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "topic_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FlashcardOut"
                  },
                  "title": "Response List Flashcards Api V1 Flashcards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Flashcards",
        "tags": [
          "practice"
        ]
      },
      "post": {
        "operationId": "create_flashcard_api_v1_flashcards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashcardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashcardOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Flashcard",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/flashcards/{card_id}": {
      "delete": {
        "operationId": "delete_flashcard_api_v1_flashcards__card_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Flashcard",
        "tags": [
          "practice"
        ]
      },
      "patch": {
        "operationId": "update_flashcard_api_v1_flashcards__card_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlashcardUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashcardOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Flashcard",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/flashcards/{card_id}/restore": {
      "post": {
        "operationId": "restore_flashcard_api_v1_flashcards__card_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlashcardOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Flashcard",
        "tags": [
          "practice"
        ]
      }
    },
...
    "/api/v1/questions": {
      "get": {
        "description": "A module's questions with your latest result on each, filtered.",
        "operationId": "question_bank_api_v1_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "topic_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          },
          {
            "in": "query",
            "name": "difficulty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "easy",
                      "medium",
                      "hard",
                      "exam"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Difficulty"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "multiple_choice",
                      "true_false",
                      "numerical",
                      "expression",
                      "short_answer",
                      "explanation",
                      "derivation"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "result",
            "required": false,
            "schema": {
              "default": "any",
              "enum": [
                "any",
                "unattempted",
                "wrong",
                "right"
              ],
              "title": "Result",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "user",
                    "claude"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "enum": [
                "active",
                "retired"
              ],
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response Question Bank Api V1 Questions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Question Bank",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/questions/{question_id}": {
      "patch": {
        "description": "Move to a topic, relabel the difficulty, or retire (and un-retire) it.\nQuestions are retired rather than deleted, so your attempts keep meaning.",
        "operationId": "update_question_api_v1_questions__question_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Question",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/quizzes": {
      "post": {
        "description": "Build a practice quiz or mock exam from the bank and start it.",
        "operationId": "create_quiz_api_v1_quizzes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Quiz",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/quizzes/{quiz_id}/attempts": {
      "post": {
        "operationId": "retake_quiz_api_v1_quizzes__quiz_id__attempts_post",
        "parameters": [
          {
            "in": "path",
            "name": "quiz_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Quiz Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retake Quiz",
        "tags": [
          "practice"
        ]
      }
    },
...
  }
}